pyproject.toml
[project]
name = "rlm-capstone"
version = "0.1.0"
description = "RLM GridWorld Experimental Framework"
requires-python = ">=3.10"

dependencies = [
    "openai>=1.0.0",
    "google-generativeai>=0.7.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v"
